backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
requests
pydantic
cloudinary
